chore(storage-resize-images): move enum
