+++ storage-resize-images/functions/lib/index.js
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.generateResizedImage = exports.deleteImage = void 0;
+exports.generateResizedImage = void 0;
 const admin = require("firebase-admin");
 const fs = require("fs");
 const functions = require("firebase-functions");
@@ -27,12 +27,6 @@ const uuidv4_1 = require("uuidv4");
 const config_1 = require("./config");
 const logs = require("./logs");
 const util_1 = require("./util");
-var deleteImage;
-(function (deleteImage) {
-    deleteImage[deleteImage["always"] = 0] = "always";
-    deleteImage[deleteImage["never"] = 1] = "never";
-    deleteImage[deleteImage["on_success"] = 2] = "on_success";
-})(deleteImage = exports.deleteImage || (exports.deleteImage = {}));
 sharp.cache(false);
 // Initialize the Firebase Admin SDK
 admin.initializeApp();
@@ -127,7 +121,7 @@ exports.generateResizedImage = functions.storage.object().onFinalize(async (obje
             fs.unlinkSync(originalFile);
             logs.tempOriginalFileDeleted(filePath);
         }
-        if (config_1.default.deleteOriginalFile === deleteImage.always) {
+        if (config_1.default.deleteOriginalFile === config_1.deleteImage.always) {
             // Delete the original file
             if (remoteFile) {
                 try {
@@ -201,7 +195,7 @@ const resizeImage = async ({ bucket, originalFile, fileDir, fileNameWithoutExten
             metadata,
         });
         logs.imageUploaded(resizedFilePath);
-        if (config_1.default.deleteOriginalFile === deleteImage.on_success) {
+        if (config_1.default.deleteOriginalFile === config_1.deleteImage.onSuccess) {
             if (remoteFile) {
                 try {
                     logs.remoteFileDeleting(filePath);
